requirements.txt
apipkg==1.5
Appium-Python-Client==0.50
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
execnet==1.7.1
idna==2.9
importlib-metadata==1.5.0
marshmallow==3.5.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-forked==1.1.3
pytest-html==2.0.1
pytest-jira==0.3.15
pytest-metadata==1.8.0
pytest-rerunfailures==8.0
pytest-testrail-e2e==2.0.0
pytest-xdist==1.31.0
requests==2.23.0
retry==0.9.2
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
zipp==3.0.0
